--- src/main/java/LunaBot/command/Command.java
@@ -1,11 +1,0 @@
-package LunaBot.command;
-
-import LunaBot.exception.LunaBotException;
-import LunaBot.storage.Storage;
-import LunaBot.task.TaskList;
-import LunaBot.ui.Ui;
-
-public abstract class Command {
-    public abstract void execute(TaskList taskList, Ui ui, Storage storage) throws LunaBotException;
-
-}
