+++ src/main/java/lunabot/command/AddEventCommand.java
@@ -1,14 +1,14 @@
-package LunaBot.command;
-
-import LunaBot.exception.LunaBotException;
-import LunaBot.storage.Storage;
-import LunaBot.task.Event;
-import LunaBot.task.TaskList;
-import LunaBot.ui.Ui;
+package lunabot.command;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+
+import lunabot.exception.LunaBotException;
+import lunabot.storage.Storage;
+import lunabot.task.Event;
+import lunabot.task.TaskList;
+import lunabot.ui.Ui;
 
 public class AddEventCommand extends Command {
     private String[] parts;
@@ -35,8 +35,7 @@ public class AddEventCommand extends Command {
             DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
             this.from = LocalDateTime.parse(parts[1], formatter);
             this.to = LocalDateTime.parse(parts[2], formatter);
-        }
-        catch  (DateTimeParseException e) {
+        } catch (DateTimeParseException e) {
             throw new LunaBotException("Invalid event date/time format. Use yyyy-MM-dd HH:mm");
         }
     }
