+++ src/main/java/lunabot/command/Command.java
@@ -1,0 +1,11 @@
+package lunabot.command;
+
+import lunabot.exception.LunaBotException;
+import lunabot.storage.Storage;
+import lunabot.task.TaskList;
+import lunabot.ui.Ui;
+
+public abstract class Command {
+    public abstract void execute(TaskList taskList, Ui ui, Storage storage) throws LunaBotException;
+
+}
