+++ src/main/java/lunabot/exception/LunaBotException.java
@@ -1,4 +1,4 @@
-package LunaBot.exception;
+package lunabot.exception;
 
 public class LunaBotException extends Exception {
     public LunaBotException(String message) {
