+++ src/main/java/lunabot/task/Event.java
@@ -1,11 +1,11 @@
-package LunaBot.task;
+package lunabot.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 public class Event extends Task {
 
-    private  LocalDateTime from;
+    private LocalDateTime from;
     private LocalDateTime to;
 
     public Event(String description, LocalDateTime from, LocalDateTime to, boolean isDone) {
@@ -27,6 +27,6 @@ public class Event extends Task {
     @Override
     public String toString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
-        return "[E]" + super.toString() + " (from: " + from.format(formatter) + " to: " + to.format(formatter) +")";
+        return "[E]" + super.toString() + " (from: " + from.format(formatter) + " to: " + to.format(formatter) + ")";
     }
 }
