+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ___________________________________________________________________
- Hello! I'm LunaBot.LunaBot
+ Hello! I'm lunabot.lunabot
  What can I do for you?
 ___________________________________________________________________
 ___________________________________________________________________
